pyproject.toml
[tool.poetry]
name = "viralmarketingproject"
version = "0.1.0"
description = ""
authors = ["kickcik <4778009@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.7"
djangorestframework = "^3.16.1"
pymysql = "^1.1.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.11"
django-filter = "^25.2"
psycopg = "^3.2.10"
djangorestframework-simplejwt = "^5.5.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.0"
black = "^25.9.0"
pre-commit = "^4.3.0"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
select = ["E", "F", "I"]  # 예: 에러, 포맷 문제, import 순서 검사
exclude = ["**/migrations/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
